refactor(benchmarks): extract finite product targets
